main.dart and auth.dart incorporated into proxy.dart
