+++ frontend/lendme/lib/screens/proxy.dart
@@ -2,13 +2,17 @@ import 'package:flutter/material.dart';
 import 'package:lendme/exceptions/general.dart';
 import 'package:lendme/models/resource.dart';
 import 'package:lendme/models/user.dart';
+import 'package:lendme/routes/auth_routes.dart';
+import 'package:lendme/routes/main_routes.dart';
 import 'package:lendme/screens/settings/edit_profile.dart';
 import 'package:provider/provider.dart';
 
-import 'auth.dart';
-import 'main.dart';
 import 'other/error.dart';
 import 'other/splash.dart';
+
+enum _Screen {
+  splash, auth, main, fillProfile, error
+}
 
 class Proxy extends StatefulWidget {
   const Proxy({Key? key}) : super(key: key);
@@ -24,8 +28,8 @@ class _ProxyState extends State<Proxy> {
 
   final Map<_Screen, Widget> _screenViews = {
     _Screen.splash: const Splash(),
-    _Screen.auth: Auth(),
-    _Screen.main: Main(),
+    _Screen.auth: PreMadeNavigator(routes: authRoutes, key: UniqueKey(),),
+    _Screen.main: PreMadeNavigator(routes: mainRoutes, key: UniqueKey()),
     _Screen.fillProfile: const EditProfile(afterLoginVariant: true),
     _Screen.error: const ErrorScreen(),
   };
@@ -85,6 +89,25 @@ class _ProxyState extends State<Proxy> {
   }
 }
 
-enum _Screen {
-  splash, auth, main, fillProfile, error
+class PreMadeNavigator extends StatelessWidget {
+  PreMadeNavigator({required this.routes, this.initialRoute='/', Key? key}) : super(key: key);
+
+  final GlobalKey<NavigatorState> _navigator = GlobalKey();
+  final String initialRoute;
+  final Route<dynamic>? Function(RouteSettings) routes;
+
+  @override
+  Widget build(BuildContext context) {
+    return WillPopScope(
+      onWillPop: () async {
+        _navigator.currentState?.maybePop();
+        return false;
+      },
+      child: Navigator(
+        key: _navigator,
+        initialRoute: initialRoute,
+        onGenerateRoute: routes,
+      ),
+    );
+  }
 }
